+++ apps/imadm/src/adm_passport_logic.erl
@@ -4,7 +4,8 @@
 % adm_passport business logic module
 %%%
 
--export([demo/3]).
+
+-export([do_login/2]).
 
 -ifdef(EUNIT).
 -include_lib("eunit/include/eunit.hrl").
@@ -13,23 +14,71 @@
 -include_lib("kernel/include/logger.hrl").
 -include_lib("imlib/include/common.hrl").
 
+
+% login_success_transfer 按顺序匹配了
+% 密码校验取第4位为密码数据 lists:nth(5, Row)
+-define (LOGIN_COLUMN, <<"id,account,mobile,password,nickname,avatar,role_id,email">>).
+
 %% ===================================================================
 %% API
 %% ===================================================================
 
 
-%%% demo方法描述
--spec demo(Uid :: integer(), Val1 :: binary(), Val2 :: binary()) -> ok.
-demo(Uid, Val1, Val2) ->
-    adm_passport_repo:demo(Uid, Val1, Val2),
-    ok.
+%%% 运营管理员登录
+
+-spec do_login(binary(), binary()) -> {ok, any()} | {error, any()}.
+do_login(_Email, <<>>) ->
+    {error, "密码有误"};
+do_login(Mobile, Pwd) ->
+    Res =
+        case imboy_func:is_mobile(Mobile) of
+            true ->
+                adm_user_repo:find_by_mobile(Mobile, ?LOGIN_COLUMN);
+            false ->
+                adm_user_repo:find_by_account(Mobile, ?LOGIN_COLUMN)
+        end,
+    % ?LOG(Res),
+    {Check, User} =
+        case Res of
+            {ok, _, [Row]} when is_tuple(Row) ->
+                % ?LOG(['Pwd', Pwd]),
+                % 第四个元素为password
+                case imboy_password:verify(Pwd, element(4, Row)) of
+                    {ok, _} ->
+                        {true, Row};
+                    {error, Msg} ->
+                        {false, Msg}
+                end;
+            _ ->
+                % io:format("res is ~p~n", [Res]),
+                {false, []}
+        end,
+    % ?LOG(['Check', Check, "user" , User]),
+    login_success_transfer(Check, User).
+
 
 
 %% ===================================================================
 %% Internal Function Definitions
 %% ===================================================================-
 
-%
+% <<"id,account,mobile,password,nickname,avatar,role_id,email">>
+-spec login_success_transfer(boolean(), tuple()) -> {ok, map()} | {error, any()}.
+login_success_transfer(true, {Id, Account, Mobile, _, Nickname, Avatar, RoleId, Email}) ->
+    {ok, #{
+           <<"id">> => imboy_hashids:uid_encode(Id),
+           <<"mobile">> => Mobile,
+           <<"email">> => Email,
+           <<"nickname">> => Nickname,
+           <<"avatar">> => Avatar,
+           <<"account">> => Account,
+           <<"role_id">> => RoleId
+          }};
+% login_success_transfer(_, User) ->
+%     ?LOG([User]),
+%     {error, "账号或密码错误"}.
+login_success_transfer(_, _) ->
+    {error, "账号或密码错误"}.
 
 %% ===================================================================
 %% EUnit tests.