+++ apps/imlib/src/imboy_response.erl
@@ -61,4 +61,7 @@ reply_json(Code, Msg, Payload, Req) ->
 reply_json(Code, Msg, Payload, Req, Options) ->
     LPayload = [{<<"code">>, Code}, {<<"msg">>, unicode:characters_to_binary(Msg)}, {<<"payload">>, Payload}],
     Body = jsone:encode(LPayload ++ Options, [native_utf8]),
-    cowboy_req:reply(200, #{<<"content-type">> => <<"application/json; charset=utf-8">>}, Body, Req).
+    cowboy_req:reply(200
+        , #{<<"content-type">> => <<"application/json; charset=utf-8">>}
+        , Body
+        , Req).